+++ example/findMovieTheaterOrganization.js
@@ -1,10 +1,6 @@
 /**
  * 枝番号で劇場組織取得サンプル
- *
- * @ignore
  */
-
-const debug = require('debug')('sskts-api-nodejs-client:samples');
 const sasaki = require('../lib/index');
 
 async function main() {
@@ -18,7 +14,7 @@ async function main() {
         state: 'teststate'
     });
     // const credentials = await auth.refreshAccessToken();
-    // debug('credentials:', credentials);
+    // console.log('credentials:', credentials);
 
     const organization = new sasaki.service.Organization({
         endpoint: process.env.API_ENDPOINT,
@@ -29,11 +25,11 @@ async function main() {
     const movieTheater = await organization.findMovieTheaterByBranchCode({
         branchCode: '118'
     });
-    debug('movieTheater is', movieTheater);
+    console.log('movieTheater is', movieTheater);
 }
 
 main().then(() => {
-    debug('main processed.');
+    console.log('main processed.');
 }).catch((err) => {
     console.error(err);
 });