+++ example/order/search.js
@@ -1,6 +1,5 @@
 /**
  * 注文検索サンプル
- * @ignore
  */
 const moment = require('moment');
 const open = require('open');