+++ example/person/closeAccount.js
@@ -1,6 +1,5 @@
 /**
  * 口座解約サンプル
- * @ignore
  */
 const moment = require('moment');
 const open = require('open');