+++ example/person/searchMoneyTransferActions.js
@@ -1,6 +1,5 @@
 /**
  * 取引履歴検索サンプル
- * @ignore
  */
 const moment = require('moment');
 const open = require('open');