+++ example/transaction/processPlaceOrderByPecorino.js
@@ -1,6 +1,5 @@
 /**
  * 口座決済による注文プロセス
- * @ignore
  */
 const moment = require('moment');
 const open = require('open');