+++ src/index.spec.ts
@@ -1,0 +1,12 @@
+/**
+ * indexテスト
+ */
+import * as assert from 'assert';
+
+import * as client from './index';
+
+describe('import client', () => {
+    it('クライアントをオブジェクトとしてインポートできるはず', () => {
+        assert.equal(typeof client, 'object');
+    });
+});