+++ CHANGELOG.md
@@ -8,10 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- 上映イベントシリーズインターフェースを追加
+
 ### Changed
 
 - 人物インターフェースを拡張
 - @cinerino/api-abstract-clientを可能な範囲で継承するように調整
+- 予約インターフェースを価格属性を除いてChevreに統一
 
 ### Deprecated