+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 人物インターフェースを拡張
 - @cinerino/api-abstract-clientを可能な範囲で継承するように調整
 - 予約インターフェースを価格属性を除いてChevreに統一
+- 個々の上映イベントのイベントタイプを変更
 
 ### Deprecated