+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - @cinerino/api-abstract-clientを可能な範囲で継承するように調整
 - 予約インターフェースを価格属性を除いてChevreに統一
 - 個々の上映イベントのイベントタイプを変更
+- 決済方法承認アクションインターフェースを強化
 
 ### Deprecated